combined into modules directory those custom classes which are used both by 1spot and amxp testruns
